cleaning up and modularizing code for more scalability
