+++ frontend/labs-chat/src/helpers/sessionId.js
@@ -1,0 +1,4 @@
+
+const sessionId = crypto.randomUUID();
+
+export default sessionId;
