+++ frontend/labs-chat/src/pages/api/streaming-llm.js
@@ -1,32 +1,14 @@
 "use server";
-import { BedrockChat } from "@langchain/community/chat_models/bedrock";
-import { ConversationChain } from "langchain/chains";
-import { BufferMemory } from "langchain/memory";
-import { NextResponse } from "next/server";
 import { HumanMessage } from "@langchain/core/messages";
-
-import { ChatPromptTemplate } from "@langchain/core/prompts";
+import { formatBedrockDataChunk, getModel } from "@/helpers/bedrock";
 
 const fakeSleep = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-const model = new BedrockChat({
-  model: process.env.MODEL,
-  region: "us-east-1",
-  streaming: true,
-  credentials: {
-    accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-    sessionToken: process.env.AWS_SESSION_TOKEN,
-  },
-  temperature: Number(process.env.MODEL_TEMPERATURE),
-});
 
-const encoder = new TextEncoder();
 function iteratorToStream(iterator) {
   return new ReadableStream({
     async pull(controller) {
-
-      // stream returns an interator that is not 
-      // stringified or encoded. we need to do both to put them 
+      // stream returns an interator that is not
+      // stringified or encoded. we need to do both to put them
       // into a proper stream format
       const chunk = await iterator.next();
       const strChunkValue = JSON.stringify(chunk.value);
@@ -37,11 +19,12 @@ function iteratorToStream(iterator) {
       // classic fake sleep issue to keep multiple chunks
       // from coming through at the same time lawlz
       await fakeSleep(10);
-      
+
       if (done) {
         controller.close();
       } else {
-        controller.enqueue(value);
+        const cleanChunk = formatBedrockDataChunk(JSON.parse(value));
+        controller.enqueue(cleanChunk);
       }
     },
   });
@@ -58,13 +41,18 @@ export default async function handler(req, res) {
   }
 
   try {
-    const request = await req.json();    
+    const request = await req.json();
     const input = request.input;
-    const iteratorBaseChunks = await model.stream([new HumanMessage({ content: input })]);
+    const iteratorBaseChunks = await getModel().stream([
+      new HumanMessage({ content: input }),
+    ]);
     const stream = iteratorToStream(iteratorBaseChunks);
     return new Response(stream);
   } catch (error) {
     console.error("Error:", error);
-    return new Response(null, { status: 500, statusText: "Internal Server Error" });
+    return new Response(null, {
+      status: 500,
+      statusText: "Internal Server Error",
+    });
   }
 }
